cleanup 7: move code to more logical layers
