+++ internal/app/internal/container/cnt_glue.go
@@ -2,17 +2,17 @@ package container
 
 import (
 	"github.com/fe3dback/go-arch-lint/internal/services/checker"
-	"github.com/fe3dback/go-arch-lint/internal/services/code"
-	"github.com/fe3dback/go-arch-lint/internal/services/path"
+	"github.com/fe3dback/go-arch-lint/internal/services/common/path"
+	"github.com/fe3dback/go-arch-lint/internal/services/common/yaml/reference"
 	"github.com/fe3dback/go-arch-lint/internal/services/project/holder"
 	"github.com/fe3dback/go-arch-lint/internal/services/project/info"
 	"github.com/fe3dback/go-arch-lint/internal/services/project/resolver"
 	"github.com/fe3dback/go-arch-lint/internal/services/project/scanner"
+	"github.com/fe3dback/go-arch-lint/internal/services/render/code"
 	"github.com/fe3dback/go-arch-lint/internal/services/schema"
 	specassembler "github.com/fe3dback/go-arch-lint/internal/services/spec/assembler"
 	"github.com/fe3dback/go-arch-lint/internal/services/spec/decoder"
 	specvalidator "github.com/fe3dback/go-arch-lint/internal/services/spec/validator"
-	"github.com/fe3dback/go-arch-lint/internal/services/yaml/reference"
 )
 
 func (c *Container) provideSpecAssembler() *specassembler.Assembler {
