+++ internal/services/checker/checker_deepscan.go
@@ -12,7 +12,7 @@ import (
 	"github.com/fe3dback/go-arch-lint/internal/models"
 	"github.com/fe3dback/go-arch-lint/internal/models/arch"
 	"github.com/fe3dback/go-arch-lint/internal/models/common"
-	"github.com/fe3dback/go-arch-lint/internal/services/deepscan"
+	"github.com/fe3dback/go-arch-lint/internal/services/checker/deepscan"
 	"golang.org/x/sync/errgroup"
 )
