+++ internal/services/checker/deepscan/ast_utils.go
