+++ internal/services/checker/deepscan/criteria.go
