+++ internal/services/checker/deepscan/models.go
