+++ internal/services/checker/deepscan/searcher.go
