+++ internal/services/checker/deepscan/searcher_parse_methods.go
