+++ internal/services/checker/deepscan/test/project/internal/di/1_basic.go
