+++ internal/services/checker/deepscan/test/project/internal/operations/0_ast_should_work.go
