+++ internal/services/checker/deepscan/test/project/internal/operations/3_receiver.go
