+++ internal/services/checker/deepscan/test/project/internal/operations/4_chan.go
