+++ internal/services/checker/deepscan/test/project/internal/operations/5_maps.go
