+++ internal/services/checker/deepscan/test/project/internal/operations/6_shared.go
