+++ internal/services/checker/deepscan/test/project/internal/operations/7_anon_interface.go
