+++ internal/services/checker/deepscan/test/project/internal/shared/interfaces.go
