+++ internal/services/checker/deepscan/utils.go
