+++ internal/services/render/code/line_count.go
