+++ internal/services/render/printer/color_printer.go
